+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to the "sfk-studio" extension will be documented in this file.
 
+## [Unreleased]
+
+### Added
+- [TypeScript]: Compile with SFK transform on saving
+- [TypeScript]: Generate Script Libraries Declaration files
+- [TypeScript]: Compile all on extension activation
+
+### Fixed
+- [TypeScript]: Rename/Move compiled JS file on TS file renaming/moving
+
 ## [1.1.1]
 
 ### Fixed